Add optional preferences.ini file & cleanup some code
Can add a [Window Settings] section to the preferences.ini and
set X, Y, Width, Height and Maximized. The window will load these
when it opens. Currently doesn't save any changes you make to the
window while it is running.
